New M5Stack_NTP_Clock_w_sleep
Modified version of the branch M5Stack_NTP_Clock.
In this version the display (and Power LED) are switched OFF and ON by touching the display. The clock part continues while display and Power LED are off. In this version only the BtnA still has a function (re-synchronize with NTP server)